app/src/main/java/com/example/myapplication/repository/TaskRepository.kt
package com.example.myapplication.repository

import androidx.lifecycle.LiveData
import com.example.myapplication.dao.TaskDao
import com.example.myapplication.model.Task

class TaskRepository(private val taskDao: TaskDao) {

    val allTasks: LiveData<List<Task>> = taskDao.getAllTasks()

    suspend fun addTask(task: Task) {
        taskDao.addTask(task)
    }

    suspend fun updateTask(task: Task) {
        taskDao.updateTask(task)
    }

    suspend fun deleteTask(task: Task) {
        taskDao.deleteTask(task)
    }
}
